generation: fix corba support
CORBA support is now disabled by default. the --[no-]corba flag to orogen
allows to enable it for a given generation. The disable_corba and enable_corba
calls to Component, Toolkit and StaticDeployment then allow to override that flag.

By default, Toolkit and StaticDeployment objects inherit the setting of their
Component