generation: make the interface more object-oriented
The current interface was relying on #initialize calls to instance_eval
their given block. In the new interface, the blocks are evaluated after
construction and #generate is called explicitely when needed. This
allows better code reuse and actually ensures that the blocks are
evaluated after everything has been properly initialized.