generation: fix CMake linking policy
CMake 2.6 asks the user to do "the right thing" for link path. I.e. it
actually expects that all libraries are given with their full path.

For now, the CMake code used in the component templates does not do
that. It was setting the related policy to NEW to avoid CMake warning,
but was not fixing the linking path. The right value was therefore OLD.