generation: fix Orocos XML marshalling for nested arrays
The current templates were reusing the same array index, leading to a scoping
problem (the inside index reinitializing the outside loop). This commit
temporarily fixes this by using C++ scoping rules (i.e. defining the index
local to the loop). A better fix would be to change the index names.