provide support for the data logger
This commit provides a ORO_UNTYPED_PROTOCOL_ID which is a fake transport whose
createBlob method returns a memory buffer with the value memcpi-ed in it. This
transport is used by the data logger to dump binary data on file.

Moreover, it changes the type names returned by orogen-generated TypeInfo objects
from the C++ style (A::B::ClassName) to the Typelib style (/A/B/ClassName) to make
it easy to use for typelib-enabled code.