template: mention that static_deployer.cmake gets rendered into OrocosComponent.cmake
When CORBA support is enabled, static_deployer.cmake is rendered at the
bottom of OrocosComponent.cmake. Thus, even though the
static_deployer.cmake file +is+ installed in the generated component,
it is not used.

This is important for buildsystem debugging (i.e. people should change
OrocosComponent.cmake instead of static_deployer.cmake)