support using opaque types in other types
This commit adds support for adding opaque types in other compounds.
The following constraints apply:

The code generation is done in this order:
 - first intermediate types
 - then opaque types
 - and finally the rest
Obviously, there is a problem when the intermediate types
depend on the other types. Moreover, you can't have an
intermediate type defined based on an opaque.

For now, the marshalling/unmarshalling is not supported.