implement marshalling of opaque fields by generating an intermediate structure
This commit solves the problem of marshalling structs that contain
an opaque type by generating another structure (with _m appended
to the original name) where opaque fields have been replaced by
their equivalent.

This removes any need for alignment and size specifications -- the
pocosim marshalling was the only place where it was used.