test: actually generate and test full components
This commit adds some orogen components, that are generated, built and ran for
testing that both the generation compiles *and* the actual functionality.

In the test_deployment case, no corba generations are used. This is for two
reasons: first, they are not really needed as these tests do not check IPC
communication. Second, OmniORB actually throws when exit() is used (which is IMO
more a problem with the way the test components are designed, so WONTFIX).