fix compilation when the component does not define any tasks
In the CMake code, the <component>-tasks-<system> library should
be linked to the deployment only if there is actually a task
library defined in the project.

The current CMake code was using component.tasks for that.
But component.tasks returns all the tasks that are available
to the user, including external tasks (as for instance the
ones defined by the OCL). One should use component.self_tasks