fix handling of default activities
Since RTT 1.8, each TaskContext has a default activity (which
is an instance of the SequentialActivity). This commit uses
'nil' as a representation of "use Orocos-provided default
activity", and updates the rest of the code to allow
TaskContext#default_activity and TaskDeployment#activity to be
nil

Moreover, a TaskDeployment's activity is now initialized with
the default activity defined for its own task context.