test: improve the common testing support
This commit removes the TEST_SKIP_REBUILD flag and replaces
it with a TEST_DONT_CLEAN, which is much more useful.

In the old case, #build_test_component was not touching the
target directory and only running the test application (if
there was one). That was making this flag useless as it was
impossible to re-run the generation and build process on
modified data.

The new TEST_DONT_CLEAN flag will now, if set, simply avoid
deleting whatever exists in test/wc, but will still run the
build process on it.

Finally, a #install method is defined to allow test cases to
install the component in its prefix without using #in_prefix