fix cross-component dependencies
This commit fixes (and tests) the use of using_task_library
and using_toolkit.

== needs_configuration? and fixed_initial_state? ==
The #fixed_initial_state? and #needs_configuration? predicate
can be now different. The first one declares that it is not
possible to specify the task context's initial state through
its constructor, while the second one specifies that this
initial state must be PreOperational. Note that, still,
needs_configuration? implies fixed_initial_state?. But the
generated code now generate constructore that allow to specify
the initial state for classes where needs_configuration or
fixed_initial_state have not been called.

This allows to derive a components for which needs_configuration?
is true from one that does not specify an initial state.

== task attributes and subclassing ==

Task object accessors (ports, methods, commands, properties and
attributes) are now in two kinds: the all_ version (all_ports) and
the self_ version (self_ports).

The all_* methods return the set of objects of that kind that exist
for an instance of that context, thus including the ones that are
defined in the parent classes.

The self_* methods return only the objects that are defined at that
particular level of the class hierarchy.