cleanup management of imported toolkits
They are now referred to by instances of their own class,
ImportedToolkit. The components#used_toolkits is no more
an array of array, and everything is fine.

Moreover, a imported_type? predicate has been added to check
if a particular type should be added to the local toolkit or
is already part of an imported one.