fix automatic toolkit regeneration in the build system
The problem is that I don't understand how the dependencies
in add_custom_command are supposed to work. Namely, I tried
having the output of the regen command to be nameTookitTypes.hpp,
assuming that would force regeneration before the tasklib is built.

But it does not work.

So we're left with having a separate stamp file. What this commit
does is ordering the regen-toolkit and the tasklib targets, so
that the regen-toolkit is called before the tasklib sources
are rebuilt (i.e. enforcing the right order).