fix #line pragma in the generated type file
the #line pragma was actually needed before the parts where we
inserted the content of a file, not before #include.

There is still an issue: the error line is getting funny if the
rest of the Types.hpp file has an error (because we have no #line
pragma to get back to the Types.hpp file).