the use of "extern template" forces to link task libraries to toolkits
This commit fixes a link bug, where port-related symbols
were not defined in the task libraries. This was caused by
the (new) usage of "extern template" in toolkits, which
speed up the compilation greatly. Caveat: one has to link
the toolkit to the task library, or the actual template
instanciation will not be found.