reimplement the toolkit generation
This is the first commit of a toolkit generation reimplementation.
The goal of the new implementation is to cleanly split toolkit and
transport (to, in fine, have one shared library for type definition
and one for transport). Moreover, it is to reduce the code size
by reusing decomposition and convertion functions in a single
toolkit and in-between toolkits.

The root of the new implementation is:
 - specific convertions functions are generated for each type
   * toPropertyBag/fromPropertyBag to (un)marshal a type to/from PropertyBag
   * toStream/fromStream to (un)marshal a type to iostream
   * toCORBA/fromCORBA to (un)marshal a type to the corresponding CORBA type
 - these methods will be reused across toolkits, improving the code size
   and possibly the speed
 - the generation of RTT type-system objects is then very easy since it
   consists only in calling these functions
 - no usage of RTT AnyConvertion anymore for the CORBA protocol. It is
   not needed anymore, and not using it has its advantages (namely, we
   can redefine types that the root RTT toolkit already defines, like
   vector<double>.
 - proper usage of ToolkitPlugin and TransportPlugin
 - clean file name usage
   * the source of a generated filr projectBlaBla.xxx is always named
     BlaBla.xxx
   * if not needed, don't split a file into multiple templates.

Finally, it will make the whole implementation manageable. The
current implementation is full of hacks and very hard to grasp.