don't try to overcome the RTT limitations at all costs
The RTT defines very little types by default. The policy until
now was to try overcoming this limitation as much as possible.
It was sometime done at the expense of correctness (for instance
casting a 64 bit integer into a 32 bit one ... no comment).

We change that. Instead, we validate method argument types and
property generation so that orogen refuses to do stuff if what
is actually missing is a type that should be defined in the RTT:
 * char, short and float are forbidden as method arguments
 * 64 bit types are forbidden as method arguments
 * the presence of a 64 bit types in a compound will forbid
   the marshalling of this compound as XML. oroGen issues a
   warning in this case.
 * char, short and float are converted into either int or double
   for XML marshalling.

Of course, these rules apply only if the types are used directly.
Using one of those is perfectly fine in a struct or vector, as
the support for them is generated by orogen.