fix major bug with handling of fixed-size arrays
From typelib's point of view, two fixed-size arrays of
different sizes are two different types.

However, from C's point of view, this is not the case: the
convertion functions are all taking T* as arguments. Compilation
was therefore failing as multiple declarations/definitions of
the same function where produced.

This commit fixes that issue by producing only one set of
convertion functions that take "T*, length" as arguments, and
importing that function in-between different toolkits.