fix overloading of methods and commands
Methods and commands are now defined as virtual methods on the
C++ side. That allows their overloading, both explicitely
(by calling method() and command() in the orogen spec) or
implicitely (just by adding the relevant C++ methods).

Additionally, it removes the build-time dependency between
the Base class and the user class.