huge cleanup of the generation of pkgconfig dependencies
This commit refactors how pkgconfig dependencies are represented
and the associated code generation is done. The rule now is to
have a #dependencies method in the relevant classes that output
a set of BuildDependency object, and then use Generation.cmake_pkgconfig*
methods to generate the CMake code.

While we are at it, reduce the number of unnecessary dependencies
and add some that were missing.