when importing task libraries, load the tasklib's own toolkit unconditionally
The issue is that it is possible to have projects that do have a toolkit
but don't have import_types_from directives, for instance because they
only define smart pointer or vector types (which are defined on the
fly).

Importing those projects led to an error because the smartptr/vector is
not regenerated when importing the tasklib. Fix that by loading the
imported project's toolkit at initialization.