fix cross-toolkit dependencies for indirect types definition
This fixes a code generation bug when a given oroGen project
(the "use" project") uses a derivative of a type that is
defined in another project (the "def" project)

This was working fine for arrays, but not for the other types.
Moreover, it was assuming that the "use" orogen file would
explicitely load the type's definition header, which is not
what we expect in practice.

Fix both issues. Cross-project dependencies for indirect types
is now supposed to work for any indirect type. It is tested for
arrays and containers.