clear() input ports on start
This is more or less an experiment, to see if it could introduce problems in the
functional layer.

It solves a common issue, in which input buffer connections get filled and
therefore do not trigger input ports anymore. If the sink task is port driven,
it does not get called at all.