rework the typelib marshaller to allow for both read and write of ports
Most of the changes have been needed to properly handle memory
management. The Handle class now tracks whether the memory pointers
it holds are owned by himself or not, and deletes the ones that are.

This is needed for proper handling of smart pointers (which own
the memory they refer to) and to allow the user to provide buffers
that should be written to ports.