modify the installed orogen file to reflect command line flags
The --extended-states and --corba flags given to the orogen command line do have
an impact on the orogen specification. Thus, it is important that, when the
project's spec is loaded, they are set to the actual values used during
generation.

We now generate a .orogen file in which these values are set to the actual
values used in generation and install the modified file.