cleanup the typelib marshaller interface
First of all, there is no need to implement readPort/writePort.
Instead, implement a method that returns a reference datasource
on the orocos sample, so that the caller can use all the operations
on datasources there is (most of them exist).

Second, no need to have const specifiers. In general, all methods
could be const, but given the API this is pretty un-informative.

cleanup setTypelibSample/refreshOrocosSample and friends to make
their semantic clearer and factor code out. Introduce
refreshTypelibSample as well.