typekit: also symlink from 'typekit' to self
This is required when the typekit is stand-alone. Since it's not in a
projectname/typekit directory, but in projectname/ directory, we don't
know in which of both the .tlb file lives. This symlink assures that we
can always access it by looking for projectname/typekit/projectname.tlb

Clearly only necessary if we use the typekit prior to installation,
like in a ros build.

Signed-off-by: Peter Soetens <peter@thesourceworks.com>