avoid a performance hit if typekit-slice is given but there is a low number of files to compile
If the number of typekit files to compile is below a given minimum,
the snippets are not grouped together to benefit from the parallel
build capabilities.

This is also useful for people that use buildd, as they can say
"don't serialize files below a threshold of 20" or something like
that.