fix duplicate code generation in typekits when both the 'all' policy is selected and some types are explicitely selected
Typekit#generate normalizes the registry, which in practice means
that @registry changes and that, therefore, all Type subclasses
change.

One effect of this is that generated_types and selected_types come
from different registries, and that they therefore cannot be gathered
in a ValueSet.

So, convert all type sets (interface, generated and selected) to the
current registry before merging them together.