in "root" typekits, generate the typelib transport for base types
The problem is that we did not have a typelib transport for base
types (int, ...) as they get registered by the RTT itself for
all other transports.

If it is not a standalone typekit (i.e. orogen and not typegen)
*and* no other typekit is being imported, generate the typelib
transport for these base types. In rock, it means that only
the orogen/base will generate it.