declare the task's destructor explicitely so that we reduce the cross-linking needs
When using default destructors, the users of the library need to link explicitely
with the SOs that define the class fields, i.e. for instance if a field is added

  class Task
  {
     Driver driver;
  };

Then a library that wants to use 'Task' needs to also link to the library defining
Driver, so as to get the default constructor and/or destructor definition.

This cross-linking requirements are removed as soon as non-default constructors/
destructors are defined. So do that for Task so as to have our users get the 'good'
behaviour by default