fix #has_opaque? in typekit generation
This fixes an issue with some builds not generating the OpaqueConvertions
files. The reason was that the only automatically generated type was
of the form std::vector<OpaqueType>, which meant that no new type needed
to be defined, and therefore the old implementation of #has_opaque? was
returning false (and OpaqueConveriton was not generated)