limit memory pressure during compilation by setting typekit_slice to 1 by default if parallel_build is greater than 4
The issue is that one TypeInfo file is ~ 400M during compilation *per type*.
What it means is that the total memory usage is 400M * typekit_slice * parallel_build

When parallel_build is 3, it means 2.4G with the default typekit slice of 2,
which starts to be a bit borderline as default on most machines. Moreover,
the typekit slice might become less interesting when compiling heavily
parallel.