spec: support loading orogen projects that are used for design purposes, not implementation
With this commit, one can give a path to an orogen file
to #using_task_library, which loads the project's task contexts
and deployments, and defines any type it adds as opaques

This can e.g. be used to design an orogen-based system without
actually creating any package, and then distributing the
orogen files.