make both the MQ and typelib targets link to the link_used_libraries from the typekit
The reason for these links is that these libraries might need
constructors / destructors which are defined by the libraries
listed in using_library. The corba transport was already fixed,
but not the other transports.

In the long run, we also should look into not linking to the typekits
from the transports (I'm not sure whether it is needed or not)