test: clear the pkgconfig cache on teardown
This fixes the TC_Component#test_imported_typekit test when the whole
test case was started in batch. The issue was that two oroGen projects
with the same name are built in different unit tests. The Utilrb::PkgConfig
caching behaviour is then making the second test fail as the orogen
project cannot be loaded (the expected paths don't exist)