fix Typekit#m_type? to return false for hand-written wrappers
We sometime write our own wrappers, that might contain m-types.
For them, Typekit#m_type? was returning true. The culprit was a
commit which made sure that the method returned true for arrays
and containers of other m-types.

Narrow the test for these types to avoid returning true on hand-written
wrapper types.