fix imported project loading
The current scheme for ImportedProject was to bypass everything to
the master project. However, this breaks e.g. the list of dependend
task libraries (basically, everyone depends on everyone with this
scheme).

This patch makes sure that imported projects load everything from their
master project, but then registers the loaded objects properly in local
sets.