gen: cleanup header handling
So far, we were generating a Types.hpp file that contains absolute
paths to all loaded headers (as well as imported typekits), and use
this header everywhere. It actually has a big performance effect on
compilation times, as well as (obviously) forcing the recompilation
of all orogen files as soon as one header is changed.

This commit implements something that got prepared for a while: generate
code that includes only the headers that are needed