Rename reuse_existing to load_lauch_file
While every launcher has to exist in the given ROS project, load_launch_file
will allow to automatically extract the defined nodes from the launch file.
Alternatively the nodes can be manually listed in the spec, e.g. when the
results of parsing are insufficient, e.g., due to includes.