Modified handling of dynamic-properties:
The old behaviour was that a dynamic property get called every time.
Now when the (renamed) __orogen_set<name> function get's called it checks
wether the task is already configured or not. If the task is not configured
then only the property get's updated. If the task was already configures
the set<name> function get's called. Therefore the user has to make sure
all properties (even the dynamic) one get's applied within the configureHook