loaders: add sanity checks
This adds checks so that projects and typekits that we load have
the name that is expected, as well as not registering twice models
with the same name(s). The latter is quite critical, as we want
to ensure that a given model in the system is uniquely represented
by a single orogen object.