loaders: auto-register children on Aggregate
The loader structure really works only if the only loader
object that is used to load models is the root. As such,
we almost always want to register the child loaders onto
their parent aggregates.

The child calls #added_child which is forwarded in Aggregate
to #add, thus allowing to override this behaviour on a case-by-case
basis.