loaders: define 'array' in the RTT typekit
The RTT typekit defines 'array' as the handler for /std/vector</double>.
It means that, in some cases, a component will report a type of 'array'
and we need to be able to resolve it.

Define here the alias of 'array' to '/std/vector</double>', which will
give access to the handlers from the RTT typekit itself. In Rock,
the 'std' typekit will export /std/vector</double> so that we get
full support for this type.