typekit.rb: fix missing typekit-name in "orogen_include"
this commit grew larger than anticipated, due to the rather convoluted
design of orogen...

as mentioned here: https://github.com/orocos-toolchain/orogen/issues/7#issuecomment-52231833

fixed to the best of my knowledge ;-) including the typekit-name in the
added line removes the doubled mention of "ReadOnlyPointer.hpp"

also renamed the "name" variable to a more precise "ptr_name".
reformulated the comments and renamed one of the "inc" variables to
"or_inc", just for a little bit extra-clarity.

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>