fix handling of containers in resolve_registry_includes (fixes #12)
The containers' source_file_line points to the template's definition, which is
then resolved to whichever header included it first. This commit adds a special
case to ensure that the metadata is properly set

It also ensures that orogen_include for a type now lists all includes that this
type requires, in particular in #ro_ptr. #ro_ptr was only adding the
ReadOnlyPointer header, not the header for the template and the header for the
element.

It removes the need for special cases for containers in include_for_type and in
template code. All code generation should now assume that orogen_include is the
authoritative source of information for includes. If it is not, the metadata
should be fixed, not the code.