test: rename all X_dir helper methods to path_to_X
This is caused by the very badly named test_dir, which is bad as
test_* are test methods. It seems that minitest/spec 5.5 actually
"hides" other test methods (and thus test_dir) from within the
other tests, which causes failures when test_dir is needed.

In any case, this commit replaces test_dir by path_to_test and
does the same for wc_root and data_dir for consistency reasons.