test: remove double-inclusion of the SelfTest modules
Starting with ruby 2.0, a module included twice in a class hierarchy
will actually be present twice. This is harmful in tests as it means
that setup and teardown are executed twice as well.

In this case, Spec is a subclass of Test and we therefore do not need
to include the SelfTest modules in both