add helpers to create "blank" project and tasks
Quite a few objects in the toolchain (in orogen, but also in e.g. syskit)
require a "null" orogen project, or an empty task context model, that is an
empty project / task context that is not linked to e.g. RTT or ROS. This
currently requires quite a bit of knowledge about the oroGen API *and* quite a
few lines of code.

Just add simple helpers to make it straightforward.