Fill empty license tag in package.xml template
This is necessary because an empty license tag is not valid. When trying
to build a package that's using this template unmodified, catkin
aborts with an error message saying that a license tag cannot be empty
or contain only whitespace. This breaks the compilation of the
ROS-Rock-bridge at the moment.