pass a component superclass at initialization time
This is a fix for a problem that plagued the plugin system for a while.
One may hook within TaskContext#initialize and e.g. create interface
elements that do not exist yet. However, if the component model subclasses
a component that already exists we get duplicates (since the subclassing
is resolved *after* #initialize).

This fixes the issue by passing the superclass at initialization time,
before super gets called