fix enabling order for default extensions in imported project
We were currently enabling default extensions at the end of the
loaded orogen project. This works in "good" cases, since any
call in user code that depends on an extension should cause
that extension to register itself (case in point: the transformer).
This user code should not be relying on the default extension
mechanism.

However, it's easy enough to make sure that the extensions are loaded
at the same point, so implement it anyways.