loaders: resolve project and typekit models on the root first
In the root/child relationship as currently implemented, models
loaded on the child are automatically registered on the root.

This fails if a system attempts to resolve a model from two
different children. For instance, Syskit has one loader per
process server, and attempts to resolve the model from a given
task/typekit on the related process server, e.g.

root: Aggregator
child1: PkgConfig (local)
child2: Orocos::RemoteProcesses::Loader

E.g. child2 is used to resolve the gps_ublox project.

However, if another component is executed on a different
process server, or if the local process needs to communicate with
components from gps_ublox, another loader will be used.

Under the current scheme, this fails as the second loader does
attempt to load the model but fails to register it on the root
(since it already exists there).

Change both project_model_from_name and typekit_model_from_name
to verify if their root already has a model first to avoid
this failure scenario.