ros: fix providing rosmap files outside of the orogen-generated paths
The loads has both a 'packs' and 'search_path' attributes. The
test sets up 'search_path', but not 'packs'

Since I can't fathom why the two are here - and can't find references
to either, I'm taking the safe route, and look for rosmap files in
both.

This fixes the test suite.