fix: do forward error(), exception() and fatal() to their generated counterparts
The previous commit forwarded these methods to the base class' version in
the setTaskState case.

This changes behavior, as currently the methods would be sending the state before
RTT changes state, while with that previous change it would be done afterwards.

Re-establish forwarding to the generated error(...), exception(...) and fatal(...)
calls to keep behavior.