fix: do not send the common state after a custom one in setTaskStates
This reestablishes the pre-setTaskState behavior, that is to not
send e.g. EXCEPTION after a custom exception in the exception/
runtime error/fatal error cases.

Since we were doing it purely from within TaskContext::exception() before,
only the "real" state was sent. Now, setTaskStates gets called *after*
the state change and this was getting published.